requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
APScheduler==3.10.4
WeasyPrint==62.3
matplotlib==3.9.0
Pillow==10.3.0
Werkzeug==3.0.3
python-dotenv==1.0.1
WTForms==3.1.2
email-validator==2.1.1
